new pipeline task for scene ID gen
