+++ convml_tt/data/sources/satellite/pipeline.py
@@ -6,18 +6,13 @@ import dateutil.parser
 import xarray as xr
 import numpy as np
 import datetime
+import isodate
 
-from . import processing, satpy_rgb, tiler, bbox
+#from . import processing, satpy_rgb, tiler, bbox
 from ....pipeline import YAMLTarget
-from ...dataset import GenericDataset
 
 SOURCE_DIR = Path("source_data")
 
-SCENE_ID_DATE_FORMAT = "%Y%m%d%H%M"
-
-
-def parse_scene_id(s):
-    return datetime.datetime.strptime(s.replace("goes16_", ""), SCENE_ID_DATE_FORMAT)
 
 
 class DatetimeListParameter(luigi.Parameter):
@@ -35,6 +30,9 @@ class GOES16Query(luigi.Task):
     debug = luigi.BoolParameter(default=False)
     data_path = luigi.Parameter()
 
+    def get_time(filename):
+        return satdata.Goes16AWS.parse_key(filename, parse_times=True)["start_time"]
+
     def run(self):
         cli = satdata.Goes16AWS(offline=False)
 
@@ -46,103 +44,28 @@ class GOES16Query(luigi.Task):
             dt_max=self.dt_max,
         )
 
-        Path(self.output().fn).parent.mkdir(exist_ok=True)
+        Path(self.output().fn).parent.mkdir(exist_ok=True, parents=True)
         self.output().write(filenames)
 
     def output(self):
-        fn = "source_data/ch{}_keys_{}.yaml".format(self.channel, self.time.isoformat())
+        fn = "ch{}_keys_{}_{}.yaml".format(self.channel, self.time.isoformat(), isodate.duration_isoformat(self.dt_max))
         p = Path(self.data_path) / fn
         return YAMLTarget(str(p))
 
 
 class GOES16Fetch(luigi.Task):
-    dt_max = luigi.FloatParameter()
-    channels = luigi.ListParameter()
-    times = DatetimeListParameter()
+    keys = luigi.ListParameter()
     data_path = luigi.Parameter()
     offline_cli = luigi.BoolParameter(default=False)
 
-    def requires(self):
-        if len(self.times) == 0:
-            raise Exception("`times` argument must have non-zero length")
-
-        reqs = {}
-        for c in self.channels:
-            reqs[c] = [
-                GOES16Query(
-                    dt_max=self.dt_max, channel=c, time=t, data_path=self.data_path
-                )
-                for t in self.times
-            ]
-        return reqs
-
     def run(self):
-        files_per_channel = {}
-        for channel, queries in self.input().items():
-            files_per_channel[channel] = []
-            for qr in queries:
-                files_per_channel[channel] += qr.read()
-
-        num_files_per_channel = {}
-        for channel, files in files_per_channel.items():
-            num_files_per_channel[channel] = len(files)
-
-        if len(set(num_files_per_channel.values())) != 1:
-            # the source data queries have resulted in a different number of
-            # files being returned for the channels selected, probably because
-            # the channels are not recorded at the same time and so one fell
-            # outside the window
-
-            def get_time(fn):
-                attrs = satdata.Goes16AWS.parse_key(fn, parse_times=True)
-                return attrs["start_time"]
-
-            def time_diff(i0, i1, fpc):
-                # fpc: new sliced files_per_channel dictionary where each list
-                # now has the same length
-                channels = list(files_per_channel.keys())
-                c1 = channels[i0]
-                c2 = channels[i1]
-                dt = get_time(fpc[c1][0]) - get_time(fpc[c2][0])
-                return abs(dt.total_seconds())
-
-            def timediff_all(fpc):
-                return sum(
-                    [time_diff(i0, i1, fpc) for (i0, i1) in [(0, 1), (0, 2), (1, 2)]]
-                )
-
-            N_max = max(num_files_per_channel.values())
-
-            fpc1 = {
-                c: len(fns) == N_max and fns[1:] or fns
-                for (c, fns) in files_per_channel.items()
-            }
-
-            fpc2 = {
-                c: len(fns) == N_max and fns[:-1] or fns
-                for (c, fns) in files_per_channel.items()
-            }
-
-            if timediff_all(fpc1) < timediff_all(fpc2):
-                files_per_channel = fpc1
-            else:
-                files_per_channel = fpc2
-
-            # now double-check that we've got the right number of files
-            num_files_per_channel = {}
-            for channel, files in files_per_channel.items():
-                num_files_per_channel[channel] = len(files)
-            assert len(set(num_files_per_channel.values())) == 1
-
         local_storage_dir = Path(self.data_path).expanduser() / SOURCE_DIR
         cli = satdata.Goes16AWS(
             offline=self.offline_cli, local_storage_dir=local_storage_dir
         )
 
-        # download all the files using the cli (flatting list first...)
-        all_files = [fn for fns in files_per_channel.values() for fn in fns]
         print("Downloading channel files...")
-        cli.download(all_files)
+        cli.download(self.keys)
 
         scenes_files = [list(a) for a in zip(*files_per_channel.values())]
 
@@ -153,16 +76,6 @@ class GOES16Fetch(luigi.Task):
 
         Path(self.output().fn).parent.mkdir(exist_ok=True)
         self.output().write(indexed_scenes_files)
-
-    def _make_scene_id(self, files):
-        attrs = satdata.Goes16AWS.parse_key(files[0], parse_times=True)
-        t = attrs["start_time"]
-        return "goes16_{}".format(t.strftime(SCENE_ID_DATE_FORMAT))
-
-    def output(self):
-        fn = "source_data/all_files.yaml"
-        p = Path(self.data_path) / fn
-        return YAMLTarget(str(p))
 
 
 class StudyTrainSplit(luigi.Task):
